Fix int casting in setFloat function

+++ Getting started/SimpleWindow/Shader.cpp
@@ -93,5 +93,5 @@ void Shader::setInt(const std::string& name, int value) const {
   glUniform1i(glGetUniformLocation(ID, name.c_str()), value);
 }
 void Shader::setFloat(const std::string& name, float value) const {
-  glUniform1f(glGetUniformLocation(ID, name.c_str()), (int)value);
+  glUniform1f(glGetUniformLocation(ID, name.c_str()), (float)value);
 }
